infrastructure/requirements.txt
aws-cdk-lib==2.150.0
constructs>=10.0.0
boto3>=1.34.0